[analyzer] Move analyzer_plugins to ClangSAConfigHandler
Move `analyzer_plugins` function and `analyzer_plugins_dir` variable
under the `ClangSAConfigHandler` class because these are only used
by this analyzer.
